FIXUP: audiotrack wrap

+++ xbmc/cores/AudioEngine/Sinks/AESinkAUDIOTRACK.h
@@ -59,7 +59,7 @@ private:
   double                m_duration_written;
   double                m_last_duration_written;
   uint64_t              m_last_head_pos;
-  uint32_t              m_head_pos_wrap_count;
+  uint64_t              m_head_pos_wrap_count;
   uint32_t              m_head_pos_reset;
 
   static std::set<unsigned int>       m_sink_sampleRates;
